02DataSpider/requirements.txt
# 信息采集模块 (Info Crawler Module) 依赖清单
# 统一UTF-8编码；已切换至Flask栈，移除FastAPI/uvicorn冲突项
# 环境标准：Python 3.11 LTS

# 爬虫框架
scrapy==2.11.0
scrapy-splash==0.8.0
playwright==1.40.0

# Telegram监听
telethon>=1.35.0
pyrogram==2.0.106

# 消息通信 - ZeroMQ
pyzmq>=25.1.2

# 数据存储
redis==5.0.1
# sqlite3 - Python内置模块，无需安装

# 数据处理 - Python 3.11/3.13兼容版本
numpy>=2.1.0,<3.0
pandas>=2.1.0,<3.0
beautifulsoup4==4.12.2
lxml>=4.9.3

# 序列化格式支持
protobuf==4.25.1
avro==1.12.0

# 配置管理
pyyaml>=6.0.2
python-dotenv==1.0.0

# 日志与监控
loguru==0.7.2
prometheus-client==0.19.0
psutil==5.9.6

# 反爬策略
fake-useragent==1.4.0
requests==2.31.0
requests-html==0.10.0

# 开发与测试
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Web API 框架
# removed: switched to Flask stack

# Flask Web API 运行栈（补充缺失依赖）
Flask==2.3.3
Flask-Cors==4.0.0
Werkzeug==2.3.7
blinker>=1.6.0
jinja2>=3.1.0
itsdangerous>=2.1.0
markupsafe>=2.1.0
click>=8.1.0

# 异步支持
aiohttp>=3.12.0
aiofiles==23.2.1

# 数据校验 - Python 3.11兼容版本
pydantic==2.5.1

# 时间处理
pytz==2023.3

# 加密与安全 - Python 3.11兼容版本
cryptography==41.0.7
